Adding pt-BR translation to the installation guide
I have a lot of brazilian friends who are struggling to get the mod to work, simply because they can't follow the installation guide easily. It would be great to include that translation into this.

I also did a "menu navigation" so people don't need to scroll down to their specific language section.